+++ CHANGES.md
@@ -1,6 +1,8 @@
 Changes 1.5.1-prerelease
 * add ChaCha20 Poly 1305 AEAD
 
+* ChaCha20: add ARMv8 NEON implementation
+
 Changes 1.5.0
 * Enable SHA3 CE 2x implementation for SLH-DSA and ML-DSA (performance increases 2 to 3 fold)